requirements.txt
django>=2.2.9,<2.3.0
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.5,<2.9.0
Faker>=4.1.2,<4.2.0

flake8>=3.6.0,<3.9.0